requirements.txt
Flask==2.3.2
Pillow==10.3.0
boto3==1.26.142
botocore==1.29.142
gunicorn==20.1.0